Fixed alarm related bugs
Fixed a bug which caused the AlarmViewer to be unable to open requested activities
Fixed a casting bug which caused the AlarmViewer not to show up

+++ app/src/main/java/com/bitflaker/lucidsourcekit/alarms/AlarmCreator.java
@@ -208,7 +208,7 @@ public class AlarmCreator extends AppCompatActivity {
         selectedToneText.setText(title);
 
         if(getIntent().hasExtra("ALARM_ID")){
-            int alarmId = getIntent().getIntExtra("ALARM_ID", -1);
+            long alarmId = getIntent().getLongExtra("ALARM_ID", -1);
             db.getStoredAlarmDao().getById(alarmId).subscribe(loadedStoredAlarm -> {
                 storedAlarm = loadedStoredAlarm;
                 setEditValuesFromItem();
